fix(settings): allow final anchors to align at the viewport top
